+++ src/gufo/thor/services/shell.py
@@ -1,0 +1,39 @@
+# ---------------------------------------------------------------------
+# Gufo Thor: shell service
+# ---------------------------------------------------------------------
+# Copyright (C) 2023, Gufo Labs
+# ---------------------------------------------------------------------
+"""
+web service.
+
+Attributes:
+    shell: shell virtual service singleton.
+"""
+
+# Gufo Thor modules
+from typing import Optional
+
+from gufo.thor.config import Config, ServiceConfig
+
+from .clickhouse import clickhouse
+from .migrate import migrate
+from .mongo import mongo
+from .noc import NocService
+from .postgres import postgres
+
+
+class ShellService(NocService):
+    """web service."""
+
+    name = "shell"
+    dependencies = (migrate, postgres, mongo, clickhouse)
+    compose_extra = {"scale": 0}
+
+    def get_compose_command(
+        self: NocService, config: Config, svc: Optional[ServiceConfig]
+    ) -> Optional[str]:
+        """Override to bash."""
+        return "/bin/bash"
+
+
+shell = ShellService()